% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_quality_check.R
\name{aof_scoring}
\alias{aof_scoring}
\title{Calculates AOF scores and scaled AOF scores}
\usage{
aof_scoring(fcs_files, phenotyping_markers, fsom, out_dir = NULL, batch = NULL)
}
\arguments{
\item{fcs_files}{Character, full path to fcs_files.}

\item{phenotyping_markers}{Character vector, marker names to be used for
clustering, can be full marker name e.g. "CD45" or "CD" if all CD-markers
needs to be plotted.}

\item{fsom}{FlowSOM object as generated by fsom_aof}

\item{out_dir}{Character, pathway to where the AOF scores and plots should
be saved, default is set to file.path(getwd(), "Quality_Control")}

\item{batch}{Character, acquisition batch for each fcs file.
This argument is passed to AOF plot names, default is set to NULL.}
}
\value{
Returns data frame with the scaled AOF scores and heatmap plots
representing AOF scores and scaled AOF scores.
}
\description{
Calculates AOF (Average Overlap Frequency) scores using flowSOM
object and greedy algorithm from cytutils package.
}
