% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_flow_rate_and_signal.R
\name{clean_files_PQC}
\alias{clean_files_PQC}
\title{Clean signal with PeacoQC}
\usage{
clean_files_PQC(
  files,
  channels_to_clean,
  cores = 1,
  plot = 20,
  out_dir = ".",
  name_dir = "Cleaned",
  MAD = 6,
  IT_limit = 0.6,
  remove_zeros = TRUE,
  arcsine_transform = TRUE,
  data_type = "MC",
  suffix_old = "beadNorm",
  suffix_new = "cleaned"
)
}
\arguments{
\item{files}{Character vector or list with the paths of the raw files.}

\item{channels_to_clean}{Channels that need to be cleaned}

\item{cores}{Number of cores to be used.}

\item{plot}{As in PeacoQC. When PeacoQC removes more than the specified
percentage, an overview plot will be made of all the selected channels and
the deleted measurements. If set to TRUE, the PlotPeacoQC function is run to
make an overview plot of the deleted measurements, even when nothing is
removed. Default is set to 20. If an increasing or decreasing trend is found,
a figure will also be made except if plot is set to FALSE.}

\item{out_dir}{As in PeacoQC. Directory where a new folder will be created
that consists of the generated fcs files, plots and report. If set to NULL,
nothing will be stored.The default folder is the working directory.}

\item{name_dir}{As in PeacoQC. Name of folder that will be generated in
\code{out_directory}.` The default is "PeacoQC_results".}

\item{MAD}{As in PeacoQC. The MAD parameter. Default is 6. If this is
increased, the algorithm becomes less strict.}

\item{IT_limit}{As in PeacoQC. The IsolationTree parameter. Default is 0.55.
If this is increased, the algorithm becomes less strict.}

\item{remove_zeros}{If this is set to TRUE, the zero values will be removed
before the peak detection step. They will not be indicated as 'bad' value.
This is recommended when cleaning mass cytometry data. Default is FALSE.}

\item{arcsine_transform}{Logical, if the data should be transformed with
arcsine transformation and cofactor 5, default is set to TRUE.
Applies to signal cleaning.}

\item{data_type}{Character, if MC (mass cytometry) of FC (flow cytometry)
data are analyzed.}

\item{suffix_old}{Old suffix to replace with \code{suffix_new}`.}

\item{suffix_new}{What to replace \code{suffix_old} with.}
}
\value{
Cleaned, untransformed flow frame if arcsine_transform argument
set to TRUE, otherwise transformed flow frame is returned.
}
\description{
Cleans the signal using PeacoQC package.
}
\examples{
# Set and create the directory where cleaned fcs files will be saved
clean_dir <- file.path(dir, "Cleaned")

# Define which files will be cleaned
files <- list.files(bead_norm_dir,
                   ignore.case = TRUE,
                   pattern = "_beadNorm.fcs$",
                   full.names = TRUE)


# Clean files
clean_files_PQC(files = files, cores = 1,
               cores = 1, remove_zeros = TRUE,
               channels_to_clean = channelsToClean,
               arcsine_transform = TRUE,
               data_type = "MC")

}
