% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{compensation}
\alias{compensation}
\title{Data Compensation}
\usage{
compensation(
  samples_input,
  compensation_input,
  choosen_channels = NULL,
  transformation = T,
  isotope_list = NULL,
  out_dir = NULL,
  suffix_old = "cleaned",
  suffix_new = "compensated",
  ...
)
}
\arguments{
\item{samples_input}{Flowframe, character vector with the paths of fcs files.}

\item{compensation_input}{Matrix, dataframe or FCS file path.}

\item{choosen_channels}{You can select channels using character vector which you used in experiment (e.g. c("Cd110Di","In115Di", "Pr141Di")).}

\item{transformation}{Default is TRUE. If you want to use transformed data, it fixes automatically according to arcsinh.(It is only for flowframe variable not fcs files.)}

\item{isotope_list}{If you have isotope which is not including in CATALYST isotope list, you can assign but it fixes automatically.}
}
\value{
returns ggplot object
}
\description{
Compensates the data using CATALYST package.
}
\examples{
samples <- list.files("./", pattern = ".fcs", full.names = T)
compensation(samples, "231115_CompBeads_Aq1_4C_01.FCS", choosen_channels = c("Cd110Di","In115Di", "Pr141Di",
                                                                             "Nd142Di", "Nd143Di", "Nd144Di",
                                                                             "Nd145Di", "Nd146Di", "Dy164Di",
                                                                             "Er167Di", "Er168Di", "Yb172Di"))

}
