% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{.gate_out_beads}
\alias{.gate_out_beads}
\title{gate_out_beads}
\usage{
.gate_out_beads(bead_channel, flow_frame)
}
\arguments{
\item{bead_channel}{character, the mass for bead channel that is exclusively used for
beads identification, no marker is present at this channel, default 140.}

\item{flow_frame}{Flow frame, if unstransformed arcsine_transform should be
kept as default, TRUE.}
}
\value{
flow frame with beads removed
}
\description{
removes beads from the files that contain them e.g files before
bead normalization
}
