% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_flowframes.R
\name{.make_breaks}
\alias{.make_breaks}
\title{Calculates breaks for flow frame splitting}
\usage{
.make_breaks(event_per_flow_frame, total_events)
}
\description{
Calculates breaks for flow frame splitting
}
