% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_flow_rate_and_signal.R
\name{.plot_flowrate}
\alias{.plot_flowrate}
\title{Plots flow rate}
\usage{
.plot_flowrate(FlowRateQC, data_type = "MC")
}
\arguments{
\item{FlowRateQC}{list obtained using flowAI:::flow_rate_check function}

\item{data_type}{Character, if "MC" (mass cytometry) or "FC" data are analyzed}
}
\value{
xgraph plot
}
\description{
plots flow rate for .fcs files
}
\references{
this code is adapted from the flowAI:::flow_rate_plot()
Monaco, G., Chen, H., Poidinger, M., Chen, J., de Magalhães, J.P.,
and Larbi, A. (2016). flowAI: automatic and interactive anomaly discerning
tools for flow cytometry data. Bioinformatics 32, 2473–2480.
}
