% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_gating.R
\name{.remove_mad_outliers}
\alias{.remove_mad_outliers}
\title{remove_mad_outliers}
\usage{
.remove_mad_outliers(
  flow_frame,
  channels = "Event_length",
  n_mad = 2,
  mad_f = mad,
  plot = TRUE,
  center = "center",
  main = "",
  ...
)
}
\arguments{
\item{flow_frame}{A flowframe that contains cytometry data.}

\item{channels}{character, channel names used for gating, default is set to
"Event_length".}

\item{n_mad}{numeric, how many MAD should be use to detect outliers}

\item{mad_f}{function used to compute deviation, default set to "mad"}

\item{plot}{logicle, if to plot the data, default TRUE}

\item{center}{the centre: defaults to the median}

\item{main}{character, title of the plot, default set to ""}

\item{...}{other arguments to pass plotDens}
}
\value{
matrix with the selected cells
}
\description{
detects outliers in the selected channel(s) using MAD
(mean absolute deviation).
}
