% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_gating.R
\name{.remove_outliers_XTPro}
\alias{.remove_outliers_XTPro}
\title{remove_outliers (i.e. multiplets)}
\usage{
.remove_outliers_XTPro(flow_frame, plot = TRUE, main = "", ...)
}
\arguments{
\item{flow_frame}{A flowframe that contains XTPRO cytometry data.}

\item{plot}{logicle, if to plot the data, default TRUE}

\item{main}{character, title of the plot, default set to ""}

\item{...}{other arguments to pass plotDens}
}
\value{
matrix with the selected cells
}
\description{
detects outliers in the selected channel(s) using MAD
(mean absolute deviation).
}
