% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_flowframes.R
\name{.split_flowFrames}
\alias{.split_flowFrames}
\title{Calculates beginning and end of each flow frame.
Code from Annelies Emmaneel (2021). PeacoQC:
Peak-based selection of high quality cytometry data. R package
version 1.4.0.}
\usage{
.split_flowFrames(vec, seg.length)
}
\description{
Calculates beginning and end of each flow frame.
Code from Annelies Emmaneel (2021). PeacoQC:
Peak-based selection of high quality cytometry data. R package
version 1.4.0.
}
