% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_quality_check.R
\name{file_outlier_detecion}
\alias{file_outlier_detecion}
\title{Detects outliers based on sample quality scores}
\usage{
file_outlier_detecion(scores, out_dir = NULL, sd)
}
\arguments{
\item{scores}{List of scaled scores per acquisition batch or data frame
of scaled scores, both generated by scaled_aof_score or aof_scoring function}

\item{out_dir}{Character, pathway to where the plot and .csv files with
quality scores should be saved, default is set to NULL, thus
file.path(getwd(), "Quality_Control" will be generated.}

\item{sd}{How many standard deviation should be use to detect outliers
default is set to 3.}
}
\value{
Save .RDS and .csv Quality scores for further analysis, and
plots and save .png for Quality AOF scores for all files.
}
\description{
Detects outlier files based on sample quality scores,
generates plot for outlier and .csv file which indicates which fcs files
could be discarded from further analysis.
}
