% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_mass_ch}
\alias{find_mass_ch}
\title{Finds all the mass/fluorochrome channels for the flow frame}
\usage{
find_mass_ch(
  flow_frame,
  channels = "Time|Event_length|Center|Offset|Width|Residual|SSC|FSC|File_scattered",
  ...
)
}
\arguments{
\item{flow_frame}{Untransformed flow frame}

\item{channels}{Pattern for non-mass channels, default is
"Time|Event_length|Center|Offset|Width|Residual|SSC|FSC|File_scattered"}

\item{...}{Additional arguments to pass to grep}
}
\value{
Logical vector with TRUE values for mass channels
}
\description{
Finds all the mass channels
}
