% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_batch_effect.R
\name{grid_arrange_common_legend}
\alias{grid_arrange_common_legend}
\title{Plot listed plots with one common legend}
\usage{
grid_arrange_common_legend(
  plot_lists,
  nrow = 1,
  ncol = length(plot_lists),
  position = c("bottom", "right")
)
}
\arguments{
\item{plot_lists}{The list of ggplots}

\item{nrow}{Numeric, number of rows to arrange the plots}

\item{ncol}{Numeric, number of columns to arrange the plots.}

\item{position}{Character, Position of the legend, either "bottom" or "right".}
}
\value{
combained ggplot
}
\description{
Plot listed plots with one common legend
}
\examples{
gg_a <- grid_arrange_common_legend(plot_lists = plots, nrow = 2, ncol = 2,
position = "right")
}
