% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{meta_data}
\alias{meta_data}
\title{Meta data containing information about sample aquisition, barcoding and
file names.}
\format{
a data frame with 33 rows and 6 columns
\describe{
\item{BARCODE}{barcode use}
\item{ID}{sample ID}
\item{STIM}{Stimulation used}
\item{DATE_EXP}{date of the experiment}
\item{DATE_AQUSITION }{date of the acqusition}
\item{BATCH}{batch id}
}
}
\usage{
meta_data
}
\description{
Meta data containing information about sample aquisition, barcoding and
file names.
}
\keyword{datasets}
