% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_2D_scatter_plots}
\alias{plot_2D_scatter_plots}
\title{Plot 2D scatter plots}
\usage{
plot_2D_scatter_plots(
  fcs_files,
  markers_to_plot,
  y_marker,
  out_dir = NULL,
  out_put_name = "marker_plots.png"
)
}
\arguments{
\item{fcs_files}{Character, pathway to fcs files.}

\item{markers_to_plot}{Character, pattern of the markers to be plotted e.g.
"CD" (all CD markers will be plotted), "CD41$" (only CD41 will be plotted).}

\item{y_marker}{Character, The marker to be plotted on the y-axis.}

\item{out_dir}{Character, path where fill are saved, if NULL (default)
files are saved in getwd()}

\item{out_put_name}{The name of the file that is saved in out_dir, default is
"marker_plots.png".}
}
\value{
Save plots in out_dir.
}
\description{
Plots biaxial plots
}
\examples{
plot_2D_scatter_plots(fcs_files = fcs_files,
                      markers_to_plot = c("CD", "HLA"),
                      y_marker = "CD45$",
                      out_dir = getwd(),
                      out_put_name = "marker_plots.png")


}
