% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{plotspillmat}
\alias{plotspillmat}
\title{Plot Spillover Matrix}
\usage{
plotspillmat(
  input,
  choosen_channels = NULL,
  transformation = T,
  isotope_list = NULL,
  savespillmat = NULL,
  ...
)
}
\arguments{
\item{input}{Matrix, dataframe or FCS file path.}

\item{choosen_channels}{You can select channels using character vector which you used in experiment (e.g. c("Cd110Di","In115Di", "Pr141Di")).}

\item{transformation}{Default is TRUE. If you want to use transformed data, it fixes automatically according to arcsinh.}

\item{isotope_list}{If you have isotope which is not including in CATALYST isotope list, you can assign but it fixes automatically.}
}
\value{
returns ggplot object
}
\description{
Creates Spillover matrix and plots.
}
\examples{
plt <- plotspillmat("231115_CompBeads_Aq1_4C_01.FCS", choosen_channels = c("Cd110Di","In115Di", "Pr141Di",
                                                                           "Nd142Di", "Nd143Di", "Nd144Di",
                                                                           "Nd145Di", "Nd146Di", "Dy164Di",
                                                                           "Er167Di", "Er168Di", "Yb172Di"))
ggsave(plt, "Spillover matrix")

}
