% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_quality_check.R
\name{scaled_aof_score}
\alias{scaled_aof_score}
\title{Calculates scaled aof scores and sample quality scores}
\usage{
scaled_aof_score(aof_scores, out_dir = NULL, aof_channels = NULL, batch = NULL)
}
\arguments{
\item{aof_scores}{Matrix, array, Aof scores obtained using function
cytutils::greedyCytometryAof.}

\item{out_dir}{Character, pathway to where the plots and scores should be
saved, if NULL file.path(getwd(), "Quality_Control) will be used.}

\item{aof_channels}{Character vector with the markers and their corresponding
channel names used for aof_scoring. Used only for plotting markers instead
of channels. If NULL (default) the colnames(aof_scores) will be used.}

\item{batch}{Character, acquisition batch for each fcs file.
Pass to FlowSOM plot name, default is set to NULL
be saved, default is set to working directory.}
}
\value{
Returns data frame with sample scores.Saved heatmaps and data frames
for AOF scores and AOF scaled scores.
}
\description{
Calculates scaled AOF scores and sample quality score.
Additionally plots heatmaps for both raw  aof scores and scaled aof scores.
}
