% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_flowframes.R
\name{split_big_flowFrames}
\alias{split_big_flowFrames}
\title{Split big flow frames into smaller fcs files}
\usage{
split_big_flowFrames(
  flow_frame,
  event_per_flow_frame = 5e+05,
  out_dir = NULL,
  min_cell_per_fcs = 20000
)
}
\arguments{
\item{flow_frame}{Flow frame containing cytometry data.}

\item{event_per_flow_frame}{Numeric, the number of events to be split to
small flow frames, default is set to 500000.}

\item{out_dir}{Character, pathway to where the files should be saved,
default is set to file.path(getwd(), Splitted).}

\item{min_cell_per_fcs}{Numeric, minimal number of cells in flow frame
to save fcs file, default 20000.}
}
\value{
Save splitted fcs files in out_dir.
}
\description{
Split the big flow frames into smaller ones.
}
\examples{
ff <- flowCore::read.FCS(list.files(path = raw_data_dir,
pattern = ".FCS", full.names = TRUE)[1])

split_big_flowFrames(flow_frame = ff, event_per_flow_frame = 100000)

}
