% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_match_order}
\alias{test_match_order}
\title{Tests the order of the files}
\usage{
test_match_order(x, y)
}
\arguments{
\item{x}{the name of the file 1 (after removing uncommon prefix/sufix)}

\item{y}{the name of the file 2 (after removing uncommon prefix/sufix)}
}
\value{
warrning if he files are not in order
}
\description{
Tests the order of the files
}
\examples{
test_match_order(x = basename(gsub("Norm_","",files_after_norm)),
basename(files_before_norm))
}
