% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_flow_rate_and_signal.R
\name{.flow_rate_bin_adapted}
\alias{.flow_rate_bin_adapted}
\title{flow_rate_bin_adapted}
\usage{
.flow_rate_bin_adapted(
  x,
  second_fraction = 0.1,
  timeCh = timeCh,
  timestep = timestep
)
}
\arguments{
\item{x}{flow frame}

\item{second_fraction}{the fraction of the seconds used in the data.}

\item{timeCh}{Time channel.}

\item{timestep}{}
}
\value{
timeFlowData, the cell assignment to the bins.
}
\description{
flow_rate_bin_adapted
}
\references{
this code is strongly based on flowAI::flow_rate_bin.
}
