% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_gating.R
\name{gate_beads}
\alias{gate_beads}
\title{Gate out beads}
\usage{
gate_beads(
  flow_frame,
  file_name = NULL,
  arcsine_transform = TRUE,
  save_gated_flow_frame = FALSE,
  out_dir = NULL,
  suffix = "_CD45_gated",
  DNA1Threshold = 5,
  ...
)
}
\arguments{
\item{flow_frame}{A flowframe that contains cytometry data.}

\item{file_name}{Character, the file name used for saving the flow frame
(if save_gated_flow_frame = TRUE) and for plotting, if NULL (default)
the file name stored in keyword FIL will be used.}

\item{arcsine_transform}{Logical, if the data should be transformed
with arcsine transformation and cofactor 5. If FALSE the data won't be
transformed, thus transformed flow frame should be used if needed.
Default TRUE.}

\item{save_gated_flow_frame}{Logical, if gated flow frame should be saved.
Only cells falling into intact cell region will be saved. Default set to FALSE.}

\item{out_dir}{Character, pathway to where the files should be saved,
if NULL (default) files will be saved to file.path(getwd(), Gated).}

\item{suffix}{Character, suffix placed in the name of saved fcs file, only
if save_gated_flow_frame = TRUE.Defult is "_intact_gated".}
}
\value{
An untransformed flow frame with non-bead events only
}
\description{
Performs gating of beads using a fixed cutoff.
}
